removed filter forms from repository file

+++ app/responses/FilterResponse.java
@@ -12,6 +12,6 @@ import java.util.List;
 public class FilterResponse<T>
 {
     private Integer offset = 0;
-    private Integer totalCount = 0;
+    private Long totalCount = 0L;
     private List<T> list = new ArrayList<>();
 }
